E2E V3 - Switched from double to int for quantity, now the REST API is communicated with

+++ src/test/java/com/optivem/atdd/e2e/v3/Params.java
@@ -18,8 +18,8 @@ public class Params {
     }
 
     // TODO: VJ: Should be an int actually
-    public double getDecimal(String key, String defaultValue) {
+    public int getInteger(String key, String defaultValue) {
         var value = paramMap.getOrDefault(key, defaultValue);
-        return Double.parseDouble(value);
+        return Integer.parseInt(value);
     }
 }
